controller: Move database layer into controller/data
