+++ controller/data/events.go
@@ -1,0 +1,340 @@
+package data
+
+import (
+	"encoding/json"
+	"errors"
+	"fmt"
+	"strconv"
+	"strings"
+	"sync"
+
+	ct "github.com/flynn/flynn/controller/types"
+	"github.com/flynn/flynn/pkg/postgres"
+	"github.com/jackc/pgx"
+)
+
+type EventRepo struct {
+	db *postgres.DB
+}
+
+func NewEventRepo(db *postgres.DB) *EventRepo {
+	return &EventRepo{db: db}
+}
+
+func (r *EventRepo) ListEvents(appID string, objectTypes []string, objectID string, beforeID *int64, sinceID *int64, count int) ([]*ct.Event, error) {
+	query := "SELECT event_id, app_id, object_id, object_type, data, op, created_at FROM events"
+	var conditions []string
+	var n int
+	args := []interface{}{}
+	if beforeID != nil {
+		n++
+		conditions = append(conditions, fmt.Sprintf("event_id < $%d", n))
+		args = append(args, *beforeID)
+	}
+	if sinceID != nil {
+		n++
+		conditions = append(conditions, fmt.Sprintf("event_id > $%d", n))
+		args = append(args, *sinceID)
+	}
+	if appID != "" {
+		n++
+		conditions = append(conditions, fmt.Sprintf("app_id = $%d", n))
+		args = append(args, appID)
+	}
+	if len(objectTypes) > 0 {
+		c := "("
+		for i, typ := range objectTypes {
+			if i > 0 {
+				c += " OR "
+			}
+			n++
+			c += fmt.Sprintf("object_type = $%d", n)
+			args = append(args, typ)
+		}
+		c += ")"
+		conditions = append(conditions, c)
+	}
+	if objectID != "" {
+		n++
+		conditions = append(conditions, fmt.Sprintf("object_id = $%d", n))
+		args = append(args, objectID)
+	}
+	if len(conditions) > 0 {
+		query += " WHERE " + strings.Join(conditions, " AND ")
+	}
+	query += " ORDER BY event_id DESC"
+	if count > 0 {
+		n++
+		query += fmt.Sprintf(" LIMIT $%d", n)
+		args = append(args, count)
+	}
+	rows, err := r.db.Query(query, args...)
+	if err != nil {
+		return nil, err
+	}
+	var events []*ct.Event
+	for rows.Next() {
+		event, err := scanEvent(rows)
+		if err != nil {
+			rows.Close()
+			return nil, err
+		}
+		events = append(events, event)
+	}
+	return events, nil
+}
+
+func (r *EventRepo) GetEvent(id int64) (*ct.Event, error) {
+	row := r.db.QueryRow("event_select", id)
+	return scanEvent(row)
+}
+
+func scanEvent(s postgres.Scanner) (*ct.Event, error) {
+	var event ct.Event
+	var typ string
+	var data []byte
+	var appID *string
+	var op *string
+	err := s.Scan(&event.ID, &appID, &event.ObjectID, &typ, &data, &op, &event.CreatedAt)
+	if err != nil {
+		if err == pgx.ErrNoRows {
+			err = ErrNotFound
+		}
+		return nil, err
+	}
+	if appID != nil {
+		event.AppID = *appID
+	}
+	if data == nil {
+		data = []byte("null")
+	}
+	if op != nil {
+		event.Op = ct.EventOp(*op)
+	}
+	event.ObjectType = ct.EventType(typ)
+	event.Data = json.RawMessage(data)
+	return &event, nil
+}
+
+// ErrEventBufferOverflow is returned to clients when the in-memory event
+// buffer is full due to clients not reading events quickly enough.
+var ErrEventBufferOverflow = errors.New("event stream buffer overflow")
+
+// eventBufferSize is the amount of events to buffer in memory.
+const eventBufferSize = 1000
+
+// EventSubscriber receives events from the EventListener loop and maintains
+// it's own loop to forward those events to the Events channel.
+type EventSubscriber struct {
+	Events  chan *ct.Event
+	Err     error
+	errOnce sync.Once
+
+	l           *EventListener
+	queue       chan *ct.Event
+	appID       string
+	objectTypes []string
+	objectID    string
+
+	stop     chan struct{}
+	stopOnce sync.Once
+}
+
+// Notify filters the event based on it's type and objectID and then pushes
+// it to the event queue.
+func (e *EventSubscriber) Notify(event *ct.Event) {
+	if len(e.objectTypes) > 0 {
+		foundType := false
+		for _, typ := range e.objectTypes {
+			if typ == string(event.ObjectType) {
+				foundType = true
+				break
+			}
+		}
+		if !foundType {
+			return
+		}
+	}
+	if e.objectID != "" && e.objectID != event.ObjectID {
+		return
+	}
+	select {
+	case e.queue <- event:
+	default:
+		// Run in a goroutine to avoid deadlock with Notify
+		go e.CloseWithError(ErrEventBufferOverflow)
+	}
+}
+
+// loop pops events off the queue and sends them to the Events channel.
+func (e *EventSubscriber) loop() {
+	defer close(e.Events)
+	for {
+		select {
+		case <-e.stop:
+			return
+		case event := <-e.queue:
+			e.Events <- event
+		}
+	}
+}
+
+// Close unsubscribes from the EventListener and stops the loop.
+func (e *EventSubscriber) Close() {
+	e.l.Unsubscribe(e)
+	e.stopOnce.Do(func() { close(e.stop) })
+}
+
+// CloseWithError sets the Err field and then closes the subscriber.
+func (e *EventSubscriber) CloseWithError(err error) {
+	e.errOnce.Do(func() { e.Err = err })
+	e.Close()
+}
+
+func NewEventListener(r *EventRepo) *EventListener {
+	return &EventListener{
+		eventRepo:   r,
+		subscribers: make(map[string]map[*EventSubscriber]struct{}),
+		doneCh:      make(chan struct{}),
+	}
+}
+
+// EventListener creates a postgres Listener for events and forwards them
+// to subscribers.
+type EventListener struct {
+	eventRepo *EventRepo
+
+	subscribers map[string]map[*EventSubscriber]struct{}
+	subMtx      sync.RWMutex
+
+	closed    bool
+	closedMtx sync.RWMutex
+	doneCh    chan struct{}
+}
+
+// Subscribe creates and returns an EventSubscriber for the given app, type and object.
+// Using an empty string for appID subscribes to all apps
+func (e *EventListener) Subscribe(appID string, objectTypes []string, objectID string) (*EventSubscriber, error) {
+	e.subMtx.Lock()
+	defer e.subMtx.Unlock()
+	if e.IsClosed() {
+		return nil, errors.New("event listener closed")
+	}
+	s := &EventSubscriber{
+		Events:      make(chan *ct.Event),
+		l:           e,
+		queue:       make(chan *ct.Event, eventBufferSize),
+		stop:        make(chan struct{}),
+		appID:       appID,
+		objectTypes: objectTypes,
+		objectID:    objectID,
+	}
+	go s.loop()
+	if _, ok := e.subscribers[appID]; !ok {
+		e.subscribers[appID] = make(map[*EventSubscriber]struct{})
+	}
+	e.subscribers[appID][s] = struct{}{}
+	return s, nil
+}
+
+// Unsubscribe unsubscribes the given subscriber.
+func (e *EventListener) Unsubscribe(s *EventSubscriber) {
+	e.subMtx.Lock()
+	defer e.subMtx.Unlock()
+	if subs, ok := e.subscribers[s.appID]; ok {
+		delete(subs, s)
+		if len(subs) == 0 {
+			delete(e.subscribers, s.appID)
+		}
+	}
+}
+
+// Listen creates a postgres listener for events and starts a goroutine to
+// forward the events to subscribers.
+func (e *EventListener) Listen() error {
+	log := logger.New("fn", "EventListener.Listen")
+	listener, err := e.eventRepo.db.Listen("events", log)
+	if err != nil {
+		e.CloseWithError(err)
+		return err
+	}
+	go func() {
+		for {
+			select {
+			case n, ok := <-listener.Notify:
+				if !ok {
+					e.CloseWithError(listener.Err)
+					return
+				}
+				idApp := strings.SplitN(n.Payload, ":", 2)
+				if len(idApp) < 1 {
+					log.Error(fmt.Sprintf("invalid event notification: %q", n.Payload))
+					continue
+				}
+				id, err := strconv.ParseInt(idApp[0], 10, 64)
+				if err != nil {
+					log.Error(fmt.Sprintf("invalid event notification: %q", n.Payload), "err", err)
+					continue
+				}
+				event, err := e.eventRepo.GetEvent(id)
+				if err != nil {
+					log.Error(fmt.Sprintf("invalid event notification: %q", n.Payload), "err", err)
+					continue
+				}
+				e.Notify(event)
+			case <-e.doneCh:
+				listener.Close()
+				return
+			}
+		}
+	}()
+	return nil
+}
+
+// Notify notifies all sbscribers of the given event.
+func (e *EventListener) Notify(event *ct.Event) {
+	e.subMtx.RLock()
+	defer e.subMtx.RUnlock()
+	if subs, ok := e.subscribers[event.AppID]; ok {
+		for sub := range subs {
+			sub.Notify(event)
+		}
+	}
+	if event.AppID != "" {
+		// Ensure subscribers not filtering by app get the event
+		if subs, ok := e.subscribers[""]; ok {
+			for sub := range subs {
+				sub.Notify(event)
+			}
+		}
+	}
+}
+
+// IsClosed returns whether or not the listener is closed.
+func (e *EventListener) IsClosed() bool {
+	e.closedMtx.RLock()
+	defer e.closedMtx.RUnlock()
+	return e.closed
+}
+
+// CloseWithError marks the listener as closed and closes all subscribers
+// with the given error.
+func (e *EventListener) CloseWithError(err error) {
+	e.closedMtx.Lock()
+	if e.closed {
+		e.closedMtx.Unlock()
+		return
+	}
+	e.closed = true
+	e.closedMtx.Unlock()
+
+	e.subMtx.RLock()
+	defer e.subMtx.RUnlock()
+	subscribers := e.subscribers
+	for _, subs := range subscribers {
+		for sub := range subs {
+			go sub.CloseWithError(err)
+		}
+	}
+	close(e.doneCh)
+}
